# Revert "sql/stats: generation count was incorrectly incremented on each look up"