--- a/pkg/sql/catalog/descs/factory.go
+++ b/pkg/sql/catalog/descs/factory.go
@@ -144,19 +144,12 @@ func (cf *CollectionFactory) NewCollection(ctx context.Context, options ...Optio
 		opt(&cfg)
 	}
 	v := cf.settings.Version.ActiveVersion(ctx)
-	// If the leaseMgr  is nil then ensure we have a nil LeaseManager interface,
-	// otherwise comparisons against a nil implementation will fail.
-	var lm LeaseManager
-	lm = cf.leaseMgr
-	if cf.leaseMgr == nil {
-		lm = nil
-	}
 	return &Collection{
 		settings:                cf.settings,
 		version:                 v,
 		hydrated:                cf.hydrated,
 		virtual:                 makeVirtualDescriptors(cf.virtualSchemas),
-		leased:                  makeLeasedDescriptors(lm),
+		leased:                  makeLeasedDescriptors(cf.leaseMgr),
 		uncommitted:             makeUncommittedDescriptors(cfg.monitor),
 		uncommittedComments:     makeUncommittedComments(),
 		uncommittedZoneConfigs:  makeUncommittedZoneConfigs(),